// ant-race/src/utils/enums.ts
export enum DataStates {
    initial = "Initial",
    loading = "Loading data...",
    loaded = "Data loaded",
  }
  
  export enum RaceStates {
    notYetRun = "Not yet run",
    inProgress = "In Progress",
    allCalculated = "All Calculated",
  }
  
  export enum CalculationStates {
    notYetRun = "Not yet run",
    inProgress = "In Progress",
    allCalculated = "Calculated",
  }
  